bugfix: fix error in urbanVPRM Pscale calculation*
This commit fixes two problems:

1) change EVImax - EVI) to (EVImax - EVImin) in the denominator of the
Pscale equation. This fixes a typo in coding Hardiman et al (2017) eq.
S4.

- Filter out non-finite Pscale values. The Pscale calculation for
urbanVPRM introduced by Hardiman et al (2017) includes a difference,
which can go to zero. Therefore the Pscale can become +/- Inf. This can
drive urbanVPRM GEE to Inf. Inf is not filtered out by the na.rm=TRUE
option for sum(), so the Inf values can drive the sum of squared errors
calculation to Inf. When this happens any set of parameter values
results in an NEE value of Inf.

When DEoptim gets a value of Inf for the cost function for every set of
parameter values it tries different values until it reaches itermax.
Every set returns Inf, so DEoptim returns the last set of parameters it
tried.  DEoptim works as designed, but essentially fails silently in
this situation.

When this happens the plots of the DEoptim results appear to bounce
around randomly.

After fixing (1) the only way to get a zero in the denominator of eq. S4
is if EVImax == EVImin; that is, EVI is constant.  That seems unlikely,
but since it is possible I am filtering these out because they cause a
hard-to-diagnose error in the overall result.